pyproject.toml
[project]
name = "discord-service-gateway"
version = "0.1.0"
description = "Discord interaction endpoint for controlling n8n and other services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "pydantic-settings>=2.5.2",
    "pynacl>=1.5.0",
    "uvicorn>=0.30.6",
]

[dependency-groups]
dev = [
    "mypy>=1.11.2",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "ruff>=0.6.8",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ASYNC", "S", "B", "C4", "DTZ", "T20", "RET", "SIM", "ARG", "PTH", "ERA", "PL", "RUF"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
